feat: extract useClickOutside into custom hook
