+++ src/hooks/useClickOutside.ts
@@ -1,0 +1,21 @@
+import { useEffect } from "react";
+
+const useClickOutside = (
+  ref: React.RefObject<HTMLElement>,
+  handler: () => void
+) => {
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        handler();
+      }
+    };
+
+    document.addEventListener("click", handleClickOutside);
+    return () => {
+      document.removeEventListener("click", handleClickOutside);
+    };
+  }, [ref, handler]);
+};
+
+export default useClickOutside;
